+++ ref-farming/tests/common/actions.rs
@@ -80,6 +80,61 @@ pub(crate) fn prepair_farm(
     (farming, farm_id)
 }
 
+pub(crate) fn prepair_multi_farms(
+    root: &UserAccount, 
+    owner: &UserAccount,
+    token: &ContractAccount<TestToken>,
+    total_reward: Balance,
+    farm_count: u32,
+) -> (ContractAccount<Farming>, Vec<String>) {
+    // create farms
+    
+    let farming = deploy_farming(&root, farming_id(), owner.account_id());
+    let mut farm_ids: Vec<String> = vec![];
+
+    // register farming contract to reward token
+    call!(
+        root,
+        token.storage_deposit(Some(to_va(farming_id())), None),
+        deposit = to_yocto("1")
+    )
+    .assert_success();
+
+    for _ in 0..farm_count {
+        let out_come = call!(
+            root,
+            farming.create_simple_farm(HRSimpleFarmTerms{
+                seed_id: format!("{}@0", swap()),
+                reward_token: to_va(token.account_id()),
+                start_at: U64(0),
+                reward_per_session: to_yocto("1").into(),
+                session_interval: U64(60),
+            }, Some(U128(1000000000000000000))),
+            deposit = to_yocto("1")
+        );
+        out_come.assert_success();
+        let farm_id: String;
+        if let Value::String(farmid) = out_come.unwrap_json_value() {
+            farm_id = farmid.clone();
+        } else {
+            farm_id = String::from("N/A");
+        }
+
+        call!(
+            root,
+            token.ft_transfer_call(to_va(farming_id()), total_reward.into(), None, farm_id.clone()),
+            deposit = 1
+        )
+        .assert_success();
+
+        farm_ids.push(farm_id.clone());
+
+        println!("  Farm {} created and running at Height#{}", farm_id.clone(), root.borrow_runtime().current_block().block_height);
+    }
+    
+    (farming, farm_ids)
+}
+
 pub(crate) fn add_liqudity(
     user: &UserAccount, 
     pool: &ContractAccount<TestRef>,
